+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - About dialog with version info + links
 - Clickable file size label in bottom right corner
 - Functionality to change the length of the data (truncate/extend)
+- Context menus in process open menu to copy addresses/sizes/etc. to clipboard
 
 ### Changed